Add a MPSC channel with backpressure handling
At the core, the channel uses an atomic FIFO queue for message passing.  This
queue is used as the primary coordination primitive. In order to enforce
capacity limits and handle back pressure, a secondary FIFO queue is used to
send parked task handles.

The general idea is that the channel is created with a `buffer` size of `n`.
The channel capacity is `n + num-senders`. Each sender gets one "guaranteed"
slot to hold a message. This allows `Sender` to know for a fact that a send
will succeed *before* starting to do the actual work of sending the value.
Since most of this work is lock-free, once the work starts, it is impossible
to safely revert.

If the sender is unable to process a send operation, then the the curren task
is parked and the handle is sent on the parked task queue.

Note that the implementation guarantees that the channel capacity will never
exceed the configured limit, however there is no *strict* guarantee that the
receiver will wake up a parked task *immediately* when a slot becomes
available. However, it will almost always unpark a task when a slot becomes
available and it is *guaranteed* that a sender will be unparked when the
message that caused the sender to become parked is read out of the channel.

The steps for sending a message are roughly:

* Increment the channel message count
* If the channel is at capacity, push the task handle onto the wait queue
* Push the message onto the message queue.

The steps for receiving a message are roughly:

* Pop a message from the message queue
* Pop a task handle from the wait queue
* Decrement the channel message count.

It's important for the order of operations on lock-free structures to happen
in reverse order between the sender and receiver. This makes the message queue
the primary coordination structure and establishes the necessary
happens-before semantics required for the acquire / release semantics used by
the queue structure.

Fixes #229