Check for channel close after a park
This fixes a deadlocking race in the mpsc channel where when a receiver is
closed it could still be possible that a sender blocks forever. The race was
that right before the parked task was sent along a channel the channel was
closed. This fixes the race by checking to see if the channel is closed after
the task is sent, ensuring that we don't needlessly block.