Unify `future::Ok`, `future::Err`, and `feature::FutureResult`
For users who are writing code that conditionally uses these
combinators, it is more convenient for them to return the same type,
without boxing or forcing a more verbose invocation. For example:

```
let x = match optional_result {
    Some(result) => future::result(result),
    None => future::err("Expected a value"),
}
```

Prior to this patch, the futures would need to be boxed, or the second
branch would need to be changed to `future::result(Err("Expected a
value"))`. Now this code compiles as expected.