Fix a race in Shared where it blocks forever
A `Shared` represents multiple handles to an original future, which can have
many tasks waiting on it. The original future, however, can only have at most
one task waiting on it. If a `Shared` handle is dropped which belongs to the
task that's blocking on the inner future, then once the inner future is done
we're not guaranteed to ever be able to wake up again.

To solve this case for now this commit adds a destructor to the `Shared` future
which wakes up all waiting tasks on destruction. This has the downside of a
"thundering stampede" problem but we can't be certain which task should be woken
up, so we just wake up all of them.

Closes #304