Add a first-class `Sink::close` method
This commit adds a new method to the `Sink` trait, `close`. The intention of
this method is to model closing a sink with an indication that no more messages
will be pushed into it. This indication can then be acted upon to do blocking
and/or other asynchronous work necessary to perform a clean shutdown.

For many protocols this is then followed by some subsequent work to shut down
the sink. For example TLS has a specific shutdown handshake which is applicable
for any TLS-backed `Sink` types.

The upcoming [tokio-io] crate recently [added] a `shutdown` method to the
`AsyncWrite` trait which this method is intended to connect up with. A call to
`Sink::close` will translate to an `AsyncWrite::shutdown` call.

Ideally this method would be a required method, but in the interest of not
breaking API compatibility just yet the method is introduced as a default method
on the `Sink` trait which calls `poll_complete`. It is strongly recommended,
however, that all implementations of `Sink` will implement this method.

More documentation about the method itself can be found in the commit with the
associated documentation.