Update Sink::poll_complete docs
This commit updates the documentation on `Sink::poll_complete` in light of #379
where it is our intention to eventually rename this method to `poll_flush`.
Additionally this method is now (and always has been) simply responsible for
flushing values from a sink.

The documentation has been updated to reflect all these aspects.