Avoid using if_std! macro
Intellij-rust plugin doesn't understand that macro and simply ignores
everthing inside of it. So, because of `if_std!` macro, completion
and code navigation does not work with future-rs which makes using
of it painful.

Although, copy-paste is bad, I think it in acceptable compromise
while tooling is not perfect, at least for the most important parts
of source.