Fill out an implementation for `Task::is_current`
This is a bit of a faceted approach. Discussion with Carl revealed that the
`is_current` answer is somewhat nuanced and not particularly well defined. To
accommodate current use cases though a new method with clearer intent and a
different implementation has been added, `will_notify_current`.

The old method is intended to soon unconditionally return `false` to allow us to
remove the global id counter. The new method doesn't work with the ids at all
and instead just does pointer comparisons.