Speed up no-std support when using std
The initial implementation of no-std introduced a performance regression
when using std. This was caused by a few factors:

* Additional calls to TLS accessors.
* Addition of dynamic dispatch.
* Addition of atomic ops.
* Less inlining

Most of these additions have been removed, leaving a single relaxed
atomic load in the `task::current()` path as well as limiting the
`call_once` to when the task is entered vs. in all task functions.