Implements `sync::slot` abstraction
This is a stream that only remembers last value and also receives
updates.

The common use case for such thing is when you have a coroutine that
refreshes a value (e.g. polls DNS server for an address), and a coroutine
that uses that value in possibly blocking fashion (e.g. fetched some
data from the address). Using unbounded stream for this case means we
can have a memory leak when consumer is too slow. And using `channel(1)`
preserves the first value while what we really need is the most recent
value generated by a producer.