Register Waker with AtomicWaker instead of context
This patch changes the AtomicWaker::register API to take a `Waker`
instance instead of `&mut Context`.

The existing implementation only needed the context to get a waker
handle. Switching the API to take the waker explicitly increases the
flexibility of `AtomicWaker`. There are cases where one wishes to
register a waker that is not associated with the current context.