Integrate futures-stable into futures crate.
futures-stable provides adapters to work with futures that contain
internal self-references.

futures-stable only works on nightly. A new nightly feature is
added to futures to turn futures-stable on. The futures-stable
items are re-exported from the `stable` submodule.

Also on nightly only, Future and Stream are implemented for
PinBoxes of futures and streams.