Some more warnings and links fixed
TODO: note that paths like futures_core::task::Context actually get redirected to core::task::Context
in the redered docs. Is this desirable? If so, should we change the paths here?

Also, there is cross crate links in here. They are not going to work anytime soon. Do we put https links
in here? to here: https://rust-lang-nursery.github.io/futures-api-docs? The problem is these have a
version hardcoded in the url: 0.3.0-alpha.15 We could link to docs.rs, but currently that says:
docs.rs failed to build futures-preview-0.3.0-alpha.15 -> ok the reason seems to be that they are on
2019-04-17 which does still have futures-api unstable feature, so that should get solved.