util: Add take_until() in StreamExt
The take_until() method stops a stream once a Future passed to take_until()
resolves. This is implemented in the TakeUntil stream combinator.
Also, the future passed to take_until() can be extracted again out of TakeUntil
with the take_future() function. In that case, the stream continues uninterrupted.
The stopping Future's eventual result can be extracted with take_result().

Docs and tests included.