Avoid writes without any data in write_all_vectored (#2187)
* Fix write-all-vectored feature

In Cargo.toml the feature is called "write-all-vectored" (with "-"), in
the code it used "write_all_vectored" (with "_"), this commit fixes the
code to use the correct name.

Also fixes the example.

* Avoid writes without any data in write_all_vectored

Previously, when non-empty sequence of empty IoSlices have been provided
to write_all_vectored, the buffers would be written as is with
write_vectored, and subsequently the return value of Ok(0) would be
misinterpreted as an error.

Avoid writes without any data by advancing the buffers first. This is
analogous to behaviour of write_all which makes no attempt to write if
there is no data.

Co-authored-by: Thomas de Zeeuw <thomasdezeeuw@gmail.com>